Modify volume view to show the picked system better
